backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
python-multipart==0.0.6
openpyxl==3.11.0
python-dotenv==1.0.0
cors==1.0.1
